refactor: UI 분리
